use rule atoms
