+++ queue/src/fiskaltrust.Middleware.Localization.QueueES/ValidationFV/ReceiptValidator.cs
@@ -1,7 +1,7 @@
-using fiskaltrust.ifPOS.v2;
+﻿using fiskaltrust.ifPOS.v2;
+using fiskaltrust.Middleware.Localization.QueueES.ValidationFV.Rules;
 using fiskaltrust.Middleware.Localization.v2.Validation;
 using FluentValidation;
-using ESValidations = fiskaltrust.Middleware.Localization.v2.Validation.Rules.ES;
 
 namespace fiskaltrust.Middleware.Localization.QueueES.ValidationFV;
 
@@ -14,7 +14,7 @@ public class ReceiptValidator : MarketValidator
 
     protected override IEnumerable<IValidator<ReceiptRequest>> GetMarketValidators()
     {
-        yield return new ESValidations.ChargeItemValidations();
-        yield return new ESValidations.CustomerValidations();
+        yield return new ChargeItemValidations();
+        yield return new CustomerValidations();
     }
 }
